@@ -1,7 +1,21 @@
 {
   "name": "peerdb-search",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.2.0",
+  "type": "module",
+  "description": "A collaborative database.",
+  "license": "Apache-2.0",
+  "author": {
+    "name": "PeerDB contributors",
+    "url": "https://gitlab.com/peerdb/peerdb"
+  },
+  "engines": {
+    "node": ">=20"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://gitlab.com/peerdb/peerdb.git"
+  },
   "scripts": {
     "serve": "vite serve",
     "build": "vite build",
@@ -11,49 +25,51 @@
     "fmt": "prettier src --write",
     "test": "vitest",
     "coverage": "vitest --coverage",
-    "test-ci": "vitest run --reporter junit --reporter default --outputFile tests.xml --coverage"
+    "test-ci": "vitest run --reporter junit --reporter default --outputFile tests.xml --coverage",
+    "upgrade": "ncu -u"
   },
   "dependencies": {
-    "@headlessui/vue": "^1.7.7",
-    "@heroicons/vue": "^2.0.11",
-    "bs58": "^5.0.0",
+    "@headlessui/vue": "^1.7.19",
+    "@heroicons/vue": "^2.1.1",
+    "@tozd/identifier": "^0.2.0",
     "esm-seedrandom": "^3.0.5",
     "lodash-es": "^4.17.21",
     "nouislider": "^15.6.1",
     "uuid": "^9.0.0",
-    "vue": "^3.2.40",
-    "vue-router": "^4.0.14"
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.0"
   },
   "devDependencies": {
-    "@babel/types": "^7.19.3",
-    "@headlessui/tailwindcss": "^0.1.1",
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.7",
-    "@types/lodash-es": "^4.17.6",
-    "@types/node": "^17.0.21",
+    "@babel/types": "^7.24.0",
+    "@headlessui/tailwindcss": "^0.2.0",
+    "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.10",
+    "@types/lodash-es": "^4.17.12",
+    "@types/node": "^20.11.24",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
-    "@vitejs/plugin-vue": "^2.2.0",
-    "@vitest/coverage-c8": "^0.23.4",
-    "@vue/eslint-config-typescript": "^10.0.0",
-    "autoprefixer": "^10.4.3",
-    "eslint": "^8.11.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-vue": "^8.5.0",
-    "postcss": "^8.4.11",
-    "postcss-html": "^1.3.0",
-    "prettier": "^2.5.1",
-    "prettier-plugin-tailwindcss": "^0.1.13",
-    "stylelint": "^14.5.3",
-    "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-standard": "^25.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitest/coverage-v8": "^1.3.1",
+    "@vue/eslint-config-typescript": "^12.0.0",
+    "autoprefixer": "^10.4.17",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-vue": "^9.22.0",
+    "npm-check-updates": "^16.14.15",
+    "postcss": "^8.4.35",
+    "postcss-html": "^1.6.0",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "rollup-plugin-license": "^3.2.0",
+    "stylelint": "^16.2.1",
+    "stylelint-config-standard": "^36.0.0",
     "stylelint-config-standard-vue": "^1.0.0",
-    "stylelint-config-tailwindcss": "^0.0.3",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.8.3",
-    "vite": "^2.8.0",
-    "vitest": "^0.24.5",
-    "vue-tsc": "^0.40.13"
+    "stylelint-config-tailwindcss": "^0.0.7",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3",
+    "vite": "^5.1.4",
+    "vitest": "^1.3.1",
+    "vue-tsc": "^1.8.27"
   }
 }
