@@ -1,9 +1,10 @@
-/* global module require */
-
-const colors = require("tailwindcss/colors")
+import colors from "tailwindcss/colors"
+import forms from "@tailwindcss/forms"
+import typography from "@tailwindcss/typography"
+import headlessui from "@headlessui/tailwindcss"
 
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
   theme: {
     extend: {
@@ -12,8 +13,9 @@ module.exports = {
         secondary: colors.yellow,
         error: colors.red,
         warning: colors.yellow,
+success: colors.green,
       },
     },
   },
-  plugins: [require("@tailwindcss/forms"), require("@tailwindcss/typography"), require("@headlessui/tailwindcss")],
+  plugins: [forms, typography, headlessui],
 }
